Refactor into input render factory - this approach aligns better with our needs for fall-through rendering
