+++ packages/components/src/internal/components/editable/constants.ts
@@ -12,28 +12,7 @@ export interface CellActions {
     selectCell: (colIdx: number, rowIdx: number, selection?: SELECTION_TYPES, resetValue?: boolean) => void;
 }
 
-export const customStyles = {
-    control: provided => ({
-        ...provided,
-        minHeight: 24,
-        borderRadius: 0,
-    }),
-    valueContainer: provided => ({
-        ...provided,
-        minHeight: 24,
-        padding: '0 4px',
-    }),
-    input: provided => ({
-        ...provided,
-        margin: '0px',
-    }),
-    indicatorsContainer: provided => ({
-        ...provided,
-        minHeight: 24,
-        padding: '0 4px',
-    }),
-};
-
+// TODO: Figure out how to reincorporate this after investigating if it is still necessary
 // Styles to match form-control in bulk form
 export const customBulkStyles = {
     control: provided => ({
@@ -47,19 +26,3 @@ export const customBulkStyles = {
         color: '#555555',
     }),
 };
-
-export const customTheme = theme => ({
-    ...theme,
-    colors: {
-        ...theme.colors,
-        danger: '#D9534F',
-        primary: '#2980B9',
-        primary75: '#009BF9',
-        primary50: '#F2F9FC',
-        primary25: 'rgba(41, 128, 185, 0.1)',
-    },
-    spacing: {
-        ...theme.spacing,
-        baseUnit: 2,
-    },
-});
